apk: move servosurface to servoview
