--- support/android/apk/servoview/src/main/java/org/mozilla/servoview/ServoSurface.java
@@ -1,448 +1,0 @@
-/* -*- Mode: Java; c-basic-offset: 4; tab-width: 4; indent-tabs-mode: nil; -*-
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
-
-package org.mozilla.servoview;
-
-import android.app.Activity;
-import android.content.Context;
-import android.util.AttributeSet;
-import android.net.Uri;
-import android.opengl.GLUtils;
-import android.os.Handler;
-import android.os.Looper;
-import android.util.DisplayMetrics;
-import android.util.Log;
-import android.view.View;
-import android.view.Surface;
-import android.view.SurfaceView;
-import android.view.SurfaceHolder;
-
-import org.mozilla.servoview.JNIServo.ServoCoordinates;
-import org.mozilla.servoview.JNIServo.ServoOptions;
-import org.mozilla.servoview.Servo.Client;
-import org.mozilla.servoview.Servo.GfxCallbacks;
-import org.mozilla.servoview.Servo.RunCallback;
-
-import android.view.Choreographer;
-import android.view.GestureDetector;
-import android.view.MotionEvent;
-import android.view.ScaleGestureDetector;
-import android.widget.OverScroller;
-
-import java.util.ArrayList;
-public class ServoSurface extends SurfaceView
-                        implements  
-                        GfxCallbacks,
-                        RunCallback,
-                        Choreographer.FrameCallback,
-                        GestureDetector.OnGestureListener,
-                        ScaleGestureDetector.OnScaleGestureListener {
-    private static final String LOGTAG = "ServoSurface";
-    private GLThread mGLThread;
-    private Handler mGLLooperHandler;
-    private Surface mASurface;
-    protected Servo mServo = null;
-    private Client mClient = null;
-    private String mServoArgs;
-    private String mServoLog;
-    private String mInitialUri;
-    private Activity mActivity;
-    private GestureDetector mGestureDetector;
-    private int mLastX = 0;
-    private int mCurX = 0;
-    private int mLastY = 0;
-    private int mCurY = 0;
-    private boolean mFlinging;
-    private ScaleGestureDetector mScaleGestureDetector;
-    private OverScroller mScroller;
-
-    private boolean mZooming;
-    private float mZoomFactor = 1;
-    private boolean mRedrawing;
-    private boolean mAnimating;
-    private boolean mPaused = false;
-
-    public ServoSurface(Context context) {
-        super(context);
-        init(context);
-    }
-
-    public ServoSurface(Context context, AttributeSet attrs) {
-        super(context, attrs);
-        init(context);
-    }
-
-    private void init(Context context) {
-        mActivity = (Activity) context;
-        setFocusable(true);
-        setFocusableInTouchMode(true);
-        setClickable(true);
-        ArrayList view = new ArrayList();
-        view.add(this);
-        addTouchables(view);
-        setWillNotCacheDrawing(false);
-        initGestures(context);
-
-        mGLThread = new GLThread(mActivity, this);
-        // this requires API level 34
-        //setSurfaceLifecycle(SurfaceView.SURFACE_LIFECYCLE_FOLLOWS_ATTACHMENT);
-        getHolder().addCallback(mGLThread);
-        mGLThread.start();
-    }
-
-
-    // Scroll and click
-
-    public void setClient(Client client) {
-        mClient = client;
-    }
-
-    public void setServoArgs(String args, String log) {
-        mServoArgs = args;
-        mServoLog = log;
-    }
-
-    // RunCallback
-    public void inGLThread(Runnable r) {
-        mGLLooperHandler.post(r);
-    }
-
-    public void inUIThread(Runnable r) {
-        post(r);
-    }
-
-    // public void shutdown() {
-    //     Log.d(LOGTAG, "shutdown");
-    //     mServo.shutdown();
-    //     mServo = null;
-    //     mGLThread.shutdown();
-    //     try {
-    //         Log.d(LOGTAG, "Waiting for GL thread to shutdown");
-    //         mGLThread.join();
-    //     } catch (InterruptedException e) {
-    //         e.printStackTrace();
-    //     }
-    // }
-    //
-
-    // GfxCallbacks
-    public void flushGLBuffers() {
-    }
-
-    public void animationStateChanged(boolean animating) {
-        if (!mAnimating && animating) {
-            post(() -> startLooping());
-        }
-        mAnimating = animating;
-    }
-
-    public void makeCurrent() {
-    }
-
-
-    private void startLooping() {
-      //In case we were already drawing.
-      Choreographer.getInstance().removeFrameCallback(this);
-
-      Choreographer.getInstance().postFrameCallback(this);
-    }
-
-
-    public void doFrame(long frameTimeNanos) {
-        Log.i(LOGTAG, ":::doFrame");
-        if (!mRedrawing) {
-            mRedrawing = true;
-            mClient.onRedrawing(mRedrawing);
-        }
-
-        // 3 reasons to be here: animating or scrolling/flinging or pinching
-
-        if (mFlinging && mScroller.isFinished()) {
-            mFlinging = false;
-            mServo.scrollEnd(0, 0, mCurX, mCurY);
-        }
-
-        if (mFlinging) {
-            mScroller.computeScrollOffset();
-            mCurX = mScroller.getCurrX();
-            mCurY = mScroller.getCurrY();
-        }
-
-        int dx = mCurX - mLastX;
-        int dy = mCurY - mLastY;
-
-        mLastX = mCurX;
-        mLastY = mCurY;
-
-        boolean scrollNecessary = mFlinging && (dx != 0 || dy != 0);
-        boolean zoomNecessary = mZooming && mZoomFactor != 1;
-
-        if (scrollNecessary) {
-            mServo.scroll(dx, dy, mCurX, mCurY);
-        }
-
-        if (zoomNecessary) {
-            mServo.pinchZoom(mZoomFactor, 0, 0);
-            mZoomFactor = 1;
-        }
-
-        if (!zoomNecessary && !scrollNecessary && mAnimating) {
-            mServo.performUpdates();
-        }
-
-        if (mZooming || mFlinging || mAnimating) {
-            Choreographer.getInstance().postFrameCallback(this);
-        } else {
-            mRedrawing = false;
-            mClient.onRedrawing(mRedrawing);
-        }
-    }
-
-    // Calls from Activity
-    public void onPause() {
-        if (mServo != null) {
-            mServo.suspend(true);
-        }
-    }
-
-
-    public void onResume() {
-        if (mServo != null) {
-            mServo.suspend(false);
-        }
-    }
-
-    public void reload() {
-        mServo.reload();
-    }
-
-    public void goBack() {
-       mServo.goBack();
-    }
-
-    public void goForward() {
-        mServo.goForward();
-    }
-
-    public void stop() {
-        mServo.stop();
-    }
-
-    public void loadUri(String uri) {
-        if (mServo != null) {
-            mServo.loadUri(uri);
-        } else {
-            mInitialUri = uri;
-        }
-    }
-
-    public void loadUri(Uri uri) {
-      loadUri(uri.toString());
-    }
-
-    public void scrollStart(int dx, int dy, int x, int y) {
-        mServo.scrollStart(dx, dy, x, y);
-    }
-
-    public void scroll(int dx, int dy, int x, int y) {
-        mServo.scroll(dx, dy, x, y);
-    }
-
-    public void scrollEnd(int dx, int dy, int x, int y) {
-        mServo.scrollEnd(dx, dy, x, y);
-    }
-
-    public void click(float x, float y) {
-        mServo.click(x, y);
-    }
-
-    public boolean onFling(MotionEvent e1, MotionEvent e2, float velocityX, float velocityY) {
-        mFlinging = true;
-
-        // FIXME: magic values
-        // https://github.com/servo/servo/issues/20361
-        int mPageWidth = 80000;
-        int mPageHeight = 80000;
-        mCurX = velocityX < 0 ? mPageWidth : 0;
-        mLastX = mCurX;
-        mCurY = velocityY < 0 ? mPageHeight : 0;
-        mLastY = mCurY;
-        mScroller.fling(mCurX, mCurY, (int) velocityX, (int) velocityY, 0, mPageWidth, 0, mPageHeight);
-        mServo.scrollStart(0, 0, mCurX, mCurY);
-        startLooping();
-        return true;
-    }
-
-    public boolean onDown(MotionEvent e) {
-        mScroller.forceFinished(true);
-        return true;
-    }
-
-    @Override
-    public boolean onTouchEvent(MotionEvent e) {
-        mGestureDetector.onTouchEvent(e);
-        mScaleGestureDetector.onTouchEvent(e);
-
-        int action = e.getActionMasked();
-
-        float x = e.getX();
-        float y = e.getY();
-
-        int pointerIndex = e.getActionIndex();
-        int pointerId = e.getPointerId(pointerIndex);
-        switch (action) {
-            case (MotionEvent.ACTION_DOWN):
-            case (MotionEvent.ACTION_POINTER_DOWN):
-                //mServo.touchDown(x, y, pointerId);
-                mFlinging = false;
-                mScroller.forceFinished(true);
-                mCurX = (int) x;
-                mLastX = mCurX;
-                mCurY = (int) y;
-                mLastY = mCurY;
-                return true;
-            case (MotionEvent.ACTION_MOVE):
-                mCurX = (int) x;
-                mCurY = (int) y;
-                //mServo.touchMove(x, y, pointerId);
-                return true;
-            case (MotionEvent.ACTION_UP):
-            case (MotionEvent.ACTION_POINTER_UP):
-                //mServo.touchUp(x, y, pointerId);
-                return true;
-            case (MotionEvent.ACTION_CANCEL):
-                //mServo.touchCancel(x, y, pointerId);
-                return true;
-            default:
-                return true;
-        }
-    }
-
-    // OnGestureListener
-    public void onLongPress(MotionEvent e) {
-    }
-
-    public boolean onScroll(MotionEvent e1, MotionEvent e2, float distanceX, float distanceY) {
-        Log.i(LOGTAG, ":::::::onScroll");
-        mServo.scroll((int) -distanceX, (int) -distanceY, (int) e1.getX(), (int) e1.getY());
-        return true;
-    }
-
-    public boolean onSingleTapUp(MotionEvent e) {
-        Log.i(LOGTAG, String.format("::onSingleTapUp %1$f %2$f", e.getX(), e.getY()));
-        click(e.getX(), e.getY());
-        return false;
-    }
-
-    public void onShowPress(MotionEvent e) {
-    }
-
-    // OnScaleGestureListener
-    @Override
-    public boolean onScaleBegin(ScaleGestureDetector detector) {
-        if (mScroller.isFinished()) {
-            mZoomFactor = detector.getScaleFactor();
-            mZooming = true;
-            mServo.pinchZoomStart(mZoomFactor, 0, 0);
-            startLooping();
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    @Override
-    public boolean onScale(ScaleGestureDetector detector) {
-        mZoomFactor *= detector.getScaleFactor();
-        return true;
-    }
-
-    @Override
-    public void onScaleEnd(ScaleGestureDetector detector) {
-        mZoomFactor = detector.getScaleFactor();
-        mZooming = false;
-        mServo.pinchZoomEnd(mZoomFactor, 0, 0);
-    }
-
-    private void initGestures(Context context) {
-        mGestureDetector = new GestureDetector(context, this);
-        mScaleGestureDetector = new ScaleGestureDetector(context, this);
-        mScroller = new OverScroller(context);
-    }
-
-    class GLThread extends Thread implements SurfaceHolder.Callback {
-        private Activity mActivity;
-        private ServoSurface mSurface;
-        GLThread(Activity activity, ServoSurface surface) {
-            mActivity = activity;
-            mSurface = surface;
-        }
-
-        public void inUIThread(Runnable r) {
-            mSurface.inUIThread(r);
-        }
-
-        public void surfaceCreated(SurfaceHolder holder) {
-            Log.d(LOGTAG, "GLThread::surfaceCreated");
-
-            ServoCoordinates coords = new ServoCoordinates();
-            coords.width = mSurface.getWidth();
-            coords.height = mSurface.getHeight();
-            coords.fb_width = mSurface.getWidth();
-            coords.fb_height = mSurface.getHeight();
-
-            Surface surface = holder.getSurface();
-            ServoOptions options = new ServoOptions();
-            options.args = mSurface.mServoArgs;
-            options.coordinates = coords;
-            options.enableLogs = true;
-            options.enableSubpixelTextAntialiasing = true;
-
-            DisplayMetrics metrics = new DisplayMetrics();
-            mActivity.getWindowManager().getDefaultDisplay().getMetrics(metrics);
-            options.density = metrics.density;
-            if (mSurface.mServo == null && !mPaused) {
-                mSurface.mServo = new Servo(options, mSurface, mSurface, mClient, mActivity, surface);
-            } else {
-                mPaused = false;
-                mSurface.mServo.resumeCompositor(surface, coords);
-            }
-
-        }
-
-        public void surfaceChanged(SurfaceHolder holder, int format, int width, int height) {
-            Log.d(LOGTAG, "GLThread::surfaceChanged");
-            ServoCoordinates coords = new ServoCoordinates();
-            coords.width = width;
-            coords.height = height;
-            coords.fb_width = width;
-            coords.fb_height = height;
-
-            if (mSurface.mServo != null) {
-                mSurface.mServo.resize(coords);
-            }
-        }
-
-        public void surfaceDestroyed(SurfaceHolder holder) {
-            Log.d(LOGTAG, "GLThread::surfaceDestroyed");
-            mPaused = true;
-            mSurface.mServo.pauseCompositor();
-        }
-
-        public void shutdown() {
-            Log.d(LOGTAG, "GLThread::shutdown");
-            // mSurface.destroy();
-            mGLLooperHandler.getLooper().quitSafely();
-        }
-
-        public void run() {
-            Looper.prepare();
-
-            mGLLooperHandler = new Handler();
-
-            Looper.loop();
-        }
-    }
-}
